+++ courlan/urlstore.py
@@ -13,6 +13,7 @@ import zlib
 from collections import defaultdict, deque
 from datetime import datetime, timedelta
 from enum import Enum
+from operator import itemgetter
 from threading import Lock
 from typing import (
     Any,
@@ -48,10 +49,10 @@ class DomainEntry:
     "Class to record host-related information and URL paths."
     __slots__ = ("count", "rules", "state", "timestamp", "total", "tuples")
 
-    def __init__(self) -> None:
+    def __init__(self, state: State = State.OPEN) -> None:
         self.count: int = 0
         self.rules: Optional[RobotFileParser] = None
-        self.state: State = State.OPEN
+        self.state: State = state
         self.timestamp: Optional[Any] = None
         self.total: int = 0
         self.tuples: Deque[UrlPathTuple] = deque()
@@ -62,25 +63,39 @@ class UrlPathTuple:
     __slots__ = ("urlpath", "visited")
 
     def __init__(self, urlpath: str, visited: bool) -> None:
-        self.urlpath: str = urlpath
+        self.urlpath: bytes = urlpath.encode("utf-8")
         self.visited: bool = visited
+
+    def path(self) -> str:
+        "Get the URL path as string."
+        return self.urlpath.decode("utf-8")
 
 
 class UrlStore:
     "Defines a class to store domain-classified URLs and perform checks against it."
-    __slots__ = ("compressed", "done", "language", "strict", "urldict", "_lock")
+    __slots__ = (
+        "compressed",
+        "done",
+        "language",
+        "strict",
+        "trailing_slash",
+        "urldict",
+        "_lock",
+    )
 
     def __init__(
         self,
         compressed: bool = False,
         language: Optional[str] = None,
         strict: bool = False,
+        trailing: bool = True,
         verbose: bool = False,
     ) -> None:
         self.compressed: bool = compressed
         self.done: bool = False
         self.language: Optional[str] = language
         self.strict: bool = strict
+        self.trailing_slash: bool = trailing
         self.urldict: DefaultDict[str, DomainEntry] = defaultdict(DomainEntry)
         self._lock: Lock = Lock()
 
@@ -112,12 +127,18 @@ class UrlStore:
                 # filter
                 if (
                     self.language is not None
-                    and lang_filter(url, self.language, self.strict) is False
+                    and lang_filter(
+                        url, self.language, self.strict, self.trailing_slash
+                    )
+                    is False
                 ):
                     LOGGER.debug("Wrong language: %s", url)
                     raise ValueError
                 parsed_url = normalize_url(
-                    parsed_url, strict=self.strict, language=self.language
+                    parsed_url,
+                    strict=self.strict,
+                    language=self.language,
+                    trailing_slash=self.trailing_slash,
                 )
                 hostinfo, urlpath = get_host_and_path(parsed_url)
                 inputdict[hostinfo].append(UrlPathTuple(urlpath, visited))
@@ -133,7 +154,7 @@ class UrlStore:
         return deque()
 
     def _set_done(self) -> None:
-        if not self.done and all(self.is_exhausted_domain(d) for d in self.urldict):
+        if not self.done and all(v.state != State.OPEN for v in self.urldict.values()):
             with self._lock:
                 self.done = True
 
@@ -163,16 +184,16 @@ class UrlStore:
             if self.urldict[domain].state is State.BUSTED:
                 return
             urls = self._load_urls(domain)
-            known = {u.urlpath for u in urls}
+            known = {u.path() for u in urls}
         else:
             urls = deque()
             known = set()
 
         # check if the link or its variants are known
         if to_right is not None:
-            urls.extend(t for t in to_right if not is_known_link(t.urlpath, known))
+            urls.extend(t for t in to_right if not is_known_link(t.path(), known))
         if to_left is not None:
-            urls.extendleft(t for t in to_left if not is_known_link(t.urlpath, known))
+            urls.extendleft(t for t in to_left if not is_known_link(t.path(), known))
 
         with self._lock:
             if self.compressed:
@@ -199,14 +220,14 @@ class UrlStore:
         # init
         last_domain: Optional[str] = None
         known_paths: Dict[str, Optional[bool]] = {}
-        remaining_urls = {u: None for u in urls}
+        remaining_urls = dict.fromkeys(urls)
         # iterate
         for url in sorted(remaining_urls):
             hostinfo, urlpath = get_host_and_path(url)
             # examine domain
             if hostinfo != last_domain:
                 last_domain = hostinfo
-                known_paths = {u.urlpath: u.visited for u in self._load_urls(hostinfo)}
+                known_paths = {u.path(): u.visited for u in self._load_urls(hostinfo)}
             # run checks: case 1: the path matches, case 2: visited URL
             if urlpath in known_paths and (
                 switch == 1 or (switch == 2 and known_paths[urlpath])
@@ -214,9 +235,6 @@ class UrlStore:
                 del remaining_urls[url]
         # preserve input order
         return list(remaining_urls)
-
-    def _timestamp(self, domain: str) -> Optional[datetime]:
-        return self.urldict[domain].timestamp
 
     # ADDITIONS AND DELETIONS
 
@@ -263,8 +281,7 @@ class UrlStore:
         "Declare domains void and prune the store."
         with self._lock:
             for d in domains:
-                self.urldict[d] = DomainEntry()
-                self.urldict[d].state = State.BUSTED
+                self.urldict[d] = DomainEntry(state=State.BUSTED)
         self._set_done()
         num = gc.collect()
         LOGGER.debug("%s objects in GC after UrlStore.discard", num)
@@ -281,22 +298,17 @@ class UrlStore:
 
     def get_known_domains(self) -> List[str]:
         "Return all known domains as a list."
-        return list(self.urldict)
+        return list(self.urldict.keys())
 
     def get_unvisited_domains(self) -> List[str]:
         """Find all domains for which there are unvisited URLs
         and potentially adjust done meta-information."""
-        unvisited = []
-        if not self.done:
-            unvisited = [d for d in self.urldict if not self.is_exhausted_domain(d)]
-            if not unvisited:
-                self._set_done()
-        return unvisited
+        return [d for d, v in self.urldict.items() if v.state == State.OPEN]
 
     def is_exhausted_domain(self, domain: str) -> bool:
         "Tell if all known URLs for the website have been visited."
         if domain in self.urldict:
-            return self.urldict[domain].state in (State.ALL_VISITED, State.BUSTED)
+            return self.urldict[domain].state != State.OPEN
         return False
         # raise KeyError("website not in store")
 
@@ -308,14 +320,12 @@ class UrlStore:
 
     def find_known_urls(self, domain: str) -> List[str]:
         """Get all already known URLs for the given domain (ex. "https://example.org")."""
-        return [domain + u.urlpath for u in self._load_urls(domain)]
+        return [domain + u.path() for u in self._load_urls(domain)]
 
     def find_unvisited_urls(self, domain: str) -> List[str]:
         "Get all unvisited URLs for the given domain."
         if not self.is_exhausted_domain(domain):
-            return [
-                domain + u.urlpath for u in self._load_urls(domain) if not u.visited
-            ]
+            return [domain + u.path() for u in self._load_urls(domain) if not u.visited]
         return []
 
     def filter_unknown_urls(self, urls: List[str]) -> List[str]:
@@ -334,7 +344,7 @@ class UrlStore:
         "Check if the given URL has already been stored."
         hostinfo, urlpath = get_host_and_path(url)
         # returns False if domain or URL is new
-        return urlpath in {u.urlpath for u in self._load_urls(hostinfo)}
+        return urlpath in {u.path() for u in self._load_urls(hostinfo)}
 
     # DOWNLOADS
 
@@ -352,27 +362,39 @@ class UrlStore:
                         with self._lock:
                             self.urldict[domain].count += 1
                         self._store_urls(domain, url_tuples, timestamp=datetime.now())
-                    return domain + url.urlpath
+                    return domain + url.path()
         # nothing to draw from
         with self._lock:
             self.urldict[domain].state = State.ALL_VISITED
         self._set_done()
         return None
 
-    def get_download_urls(self, timelimit: int = 10) -> Optional[List[str]]:
+    def get_download_urls(
+        self,
+        time_limit: int = 10,
+        max_urls: int = 10000,
+        timelimit: Optional[int] = None,  # TODO: remove later
+    ) -> Optional[List[str]]:
         """Get a list of immediately downloadable URLs according to the given
         time limit per domain."""
-        potential = self.get_unvisited_domains()
-        targets = []
-        for domain in potential:
-            timestamp = self._timestamp(domain)
+        if timelimit is not None:
+            raise ValueError("timelimit is deprecated, use time_limit instead")
+
+        urls = []
+        for website, entry in self.urldict.items():
+            if entry.state != State.OPEN:
+                continue
             if (
-                timestamp is None
-                or (datetime.now() - timestamp).total_seconds() > timelimit
+                not entry.timestamp
+                or (datetime.now() - entry.timestamp).total_seconds() > time_limit
             ):
-                targets.append(domain)
-        # get corresponding URLs and filter out None values
-        return list(filter(None, [self.get_url(domain) for domain in targets]))
+                url = self.get_url(website)
+                if url is not None:
+                    urls.append(url)
+                    if len(urls) >= max_urls:
+                        break
+        self._set_done()
+        return urls
 
     def establish_download_schedule(
         self, max_urls: int = 100, time_limit: int = 10
@@ -399,15 +421,15 @@ class UrlStore:
                 ):
                     break
                 if not url.visited:
-                    urlpaths.append(url.urlpath)
+                    urlpaths.append(url.path())
                     url.visited = True
                     with self._lock:
                         self.urldict[domain].count += 1
             # determine timestamps
             now = datetime.now()
-            original_timestamp = self._timestamp(domain)
+            original_timestamp = self.urldict[domain].timestamp
             if (
-                original_timestamp is None
+                not original_timestamp
                 or (now - original_timestamp).total_seconds() > time_limit
             ):
                 schedule_secs = 0.0
@@ -424,7 +446,7 @@ class UrlStore:
             self._store_urls(domain, url_tuples, timestamp=total_diff)
         # sort by first tuple element (time in secs)
         self._set_done()
-        return sorted(targets, key=lambda x: x[0])  # type: ignore[arg-type]
+        return sorted(targets, key=itemgetter(1))  # type: ignore[arg-type]
 
     # CRAWLING
 
@@ -459,15 +481,15 @@ class UrlStore:
 
     def get_all_counts(self) -> List[int]:
         "Return all download counts for the hosts in store."
-        return [self.urldict[d].count for d in self.urldict]
+        return [v.count for v in self.urldict.values()]
 
     def total_url_number(self) -> int:
         "Find number of all URLs in store."
-        return sum(self.urldict[d].total for d in self.urldict)
+        return sum(v.total for v in self.urldict.values())
 
     def download_threshold_reached(self, threshold: float) -> bool:
         "Find out if the download limit (in seconds) has been reached for one of the websites in store."
-        return any(self.urldict[d].count >= threshold for d in self.urldict)
+        return any(v.count >= threshold for v in self.urldict.values())
 
     def dump_urls(self) -> List[str]:
         "Return a list of all known URLs."
@@ -487,9 +509,25 @@ class UrlStore:
             print(
                 "\n".join(
                     [
-                        domain + u.urlpath + "\t" + str(u.visited)
+                        f"{domain}{u.path()}\t{str(u.visited)}"
                         for u in self._load_urls(domain)
                     ]
                 ),
                 flush=True,
             )
+
+    # PERSISTANCE
+
+    def write(self, filename: str) -> None:
+        "Write the URL store to disk."
+        del self._lock
+        with open(filename, "wb") as output:
+            pickle.dump(self, output)
+
+
+def load_store(filename: str) -> UrlStore:
+    "Load a URL store from disk."
+    with open(filename, "rb") as output:
+        url_store = pickle.load(output)
+    url_store._lock = Lock()
+    return url_store  # type: ignore[no-any-return]