+++ tests/urlstore_tests.py
@@ -7,6 +7,7 @@ import os
 import pickle
 import signal
 import sys
+import tempfile
 import uuid
 
 from datetime import datetime
@@ -15,8 +16,7 @@ from time import sleep
 import pytest
 
 from courlan import UrlStore
-from courlan.core import filter_links
-from courlan.urlstore import State
+from courlan.urlstore import State, load_store
 
 
 def test_urlstore():
@@ -37,7 +37,7 @@ def test_urlstore():
     assert len(my_urls.urldict) == 1 and "http://example.org" not in my_urls.urldict
     assert len(my_urls.urldict["https://example.org"].tuples) == 2
     firstelem = my_urls.urldict["https://example.org"].tuples[0]
-    assert firstelem.urlpath == "/" and firstelem.visited is False
+    assert firstelem.urlpath == b"/" and firstelem.visited is False
     # reset
     num, _, _ = gc.get_count()
     my_urls.reset()
@@ -182,11 +182,12 @@ def test_urlstore():
     my_urls.add_urls(extension_urls)
     assert len(my_urls._load_urls(example_domain)) == len(example_urls) + 10
     # test extension + deduplication
+    my_urls.trailing_slash = False
     extension_urls = [f"{example_domain}/1/{str(a)}/" for a in range(11)]
     my_urls.add_urls(appendleft=extension_urls)
     url_tuples = my_urls._load_urls(example_domain)
     assert len(url_tuples) == len(example_urls) + 11
-    assert url_tuples[-1].urlpath == "/1/9" and url_tuples[0].urlpath == "/1/10/"
+    assert url_tuples[-1].urlpath == b"/1/9" and url_tuples[0].urlpath == b"/1/10"
 
     # duplicates
     my_urls.add_urls(extension_urls)
@@ -194,7 +195,7 @@ def test_urlstore():
     assert len(my_urls._load_urls(example_domain)) == len(example_urls) + len(
         extension_urls
     )
-    assert url_tuples[-1].urlpath == "/1/9" and url_tuples[0].urlpath == "/1/10/"
+    assert url_tuples[-1].urlpath == b"/1/9" and url_tuples[0].urlpath == b"/1/10"
 
     # get_url
     assert my_urls.urldict[example_domain].timestamp is None
@@ -204,7 +205,7 @@ def test_urlstore():
     timestamp = my_urls.urldict[example_domain].timestamp
     sleep(0.1)
     url2 = my_urls.get_url(example_domain)
-    assert url1 != url2 and url1 == "https://www.example.org/1/10/"
+    assert url1 != url2 and url1 == "https://www.example.org/1/10"
     assert my_urls.urldict[example_domain].count == 2
     assert timestamp != my_urls.urldict[example_domain].timestamp
     assert url2 not in set(my_urls.find_unvisited_urls(example_domain))
@@ -220,7 +221,9 @@ def test_urlstore():
 
     url_tuples = my_urls._load_urls(example_domain)
     # positions
-    assert url1.endswith(url_tuples[0].urlpath) and url2.endswith(url_tuples[1].urlpath)
+    assert url1.endswith(url_tuples[0].urlpath.decode("utf-8")) and url2.endswith(
+        url_tuples[1].urlpath.decode("utf-8")
+    )
     # timestamp
     assert my_urls.urldict[example_domain].timestamp is not None
     # nothing left
@@ -285,18 +288,36 @@ def test_urlstore():
     assert my_urls.total_url_number() == 20014
 
     # get download URLs
-    downloadable_urls = my_urls.get_download_urls(timelimit=0)
+    downloadable_urls = my_urls.get_download_urls(time_limit=0, max_urls=1)
     assert (
-        len(downloadable_urls) == 2
+        len(downloadable_urls) == 1
         and downloadable_urls[0] == "https://www.example.org/1"
     )
     assert (
         datetime.now() - my_urls.urldict["https://www.example.org"].timestamp
     ).total_seconds() < 0.25
     assert my_urls.urldict["https://www.example.org"].count == 3
-    assert my_urls.urldict["https://test.org"].count == 1
-    downloadable_urls = my_urls.get_download_urls()  # limit=10
+
+    # does not work on Windows?
+    # if os.name != "nt":
+    test_urls = UrlStore()
+    test_urls.add_urls(
+        ["https://www.example.org/1", "https://test.org/1", "https://test.org/2"]
+    )
+
+    with pytest.raises(ValueError):
+        test_urls.get_download_urls(timelimit=0)
+
+    downloadable_urls = test_urls.get_download_urls(time_limit=0)
+    assert (
+        len(downloadable_urls) == 2
+        and downloadable_urls[0].startswith("https://www.example.org")
+        and downloadable_urls[1].startswith("https://test.org")
+        and test_urls.urldict["https://test.org"].count == 1
+    )
+    downloadable_urls = test_urls.get_download_urls()
     assert len(downloadable_urls) == 0
+
     other_store = UrlStore()
     downloadable_urls = other_store.get_download_urls()
     assert not downloadable_urls and other_store.done is True
@@ -318,14 +339,14 @@ def test_urlstore():
     assert (
         len(schedule) == 1
         and round(schedule[0][0]) == 1
-        and schedule[0][1] == "https://www.example.org/2"
+        and schedule[0][1].startswith("https://www.example.org")
     )
     schedule = my_urls.establish_download_schedule(max_urls=6, time_limit=1)
     assert len(schedule) == 6 and round(max(s[0] for s in schedule)) == 4
     assert my_urls.urldict["https://www.example.org"].count == 7
     assert (
         my_urls.urldict["https://test.org"].count
-        == 4
+        == 3
         == sum(u.visited is True for u in my_urls.urldict["https://test.org"].tuples)
     )
     assert my_urls.download_threshold_reached(8) is False
@@ -371,6 +392,7 @@ def test_dbdump(capsys):
 def test_from_html():
     "Test link extraction procedures."
     url_store = UrlStore()
+    url_store.trailing_slash = False
     base_url = "https://example.org"
     htmlstring = '<html><body><a href="https://example.com/page1"/><a href="https://example.org/page1/"/><a href="https://test.org/page1"/></body></html>'
     # 1 internal link in total
@@ -398,7 +420,7 @@ def test_from_html():
     url_store.add_from_html(htmlstring, base_url, lang="en")
     todo = url_store.find_unvisited_urls(base_url)
     known_links = url_store.find_known_urls(base_url)
-    assert "https://example.org/en/page1/" in todo and len(known_links) == 4
+    assert "https://example.org/en/page1" in todo and len(known_links) == 4
     # wrong language
     htmlstring = '<html><body><a href="https://example.org/en/page2"/></body></html>'
     url_store.add_from_html(htmlstring, base_url, lang="de")
@@ -412,3 +434,29 @@ def test_from_html():
     htmlstring = '<html><body><a href="https://example.org/en/page2"/><a href="https://example.org/imprint.html"/></body></html>'
     url_store.add_from_html(htmlstring, base_url)
     assert not url_store.find_known_urls(base_url)
+
+
+def test_persistance():
+    "Test writing and loading to/from disk."
+    url_store = UrlStore(
+        compressed=True, language="de", strict=True, trailing=True, verbose=True
+    )
+    example_urls = [f"https://www.example.org/{str(a)}" for a in range(100)]
+    test_urls = [f"https://test.org/{str(uuid.uuid4())[:20]}" for _ in range(100)]
+    url_store.add_urls(example_urls + test_urls)
+
+    _, tmp = tempfile.mkstemp()
+    url_store.write(tmp)
+    new_store = load_store(tmp)
+    try:
+        os.remove(tmp)
+    except PermissionError:
+        pass  # Windows
+
+    assert new_store.compressed is True
+    assert new_store.language == "de"
+    assert new_store.strict is True
+    assert new_store.trailing_slash is True
+    urls = set(new_store.dump_urls())
+    assert new_store.total_url_number() == len(urls) == 200
+    assert "https://www.example.org/99" in urls